moved weekly dose inside knowledge cafe
