+++ app/Http/Requests/KnowledgeCafe/WeeklyDoseRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\KnowledgeCafe;
 
 use Illuminate\Foundation\Http\FormRequest;
