+++ app/Providers/AuthServiceProvider.php
@@ -16,7 +16,7 @@ class AuthServiceProvider extends ServiceProvider
         'App\Models\Client' => 'App\Policies\ClientPolicy',
         'App\Models\Finance\Invoice' => 'App\Policies\Finance\InvoicePolicy',
         'App\Models\Project' => 'App\Policies\ProjectPolicy',
-        'App\Models\WeeklyDose' => 'App\Policies\WeeklyDosePolicy',
+        'App\Models\KnowledgeCafe\WeeklyDose' => 'App\Policies\KnowledgeCafe\WeeklyDosePolicy',
         'App\Models\HR\Job' => 'App\Policies\HR\JobPolicy',
         'App\Models\HR\Applicant' => 'App\Policies\HR\ApplicantPolicy',
         'App\Models\KnowledgeCafe\Library\Book' => 'App\Policies\KnowledgeCafe\Library\BookPolicy',
