+++ artisan
@@ -15,9 +15,9 @@ define('LARAVEL_START', microtime(true));
 |
 */
 
-require __DIR__.'/vendor/autoload.php';
+require __DIR__.'\vendor\autoload.php';
 
-$app = require_once __DIR__.'/bootstrap/app.php';
+$app = require_once __DIR__.'\bootstrap\app.php';
 
 /*
 |--------------------------------------------------------------------------
