+++ routes/web.php
@@ -53,7 +53,7 @@ Route::middleware('auth')->group(function () {
 
     Route::resource('clients', 'ClientController')->except(['show', 'destroy']);
     Route::resource('projects', 'ProjectController')->except(['show', 'destroy']);
-    Route::resource('weeklydoses', 'WeeklyDoseController')->only(['index']);
+    // Route::resource('weeklydoses', 'WeeklyDoseController')->only(['index']);
     Route::get('clients/{client}/get-projects', 'ClientController@getProjects');
 
     Route::resource('project/stages', 'ProjectStageController')->only(['store', 'update']);
@@ -62,7 +62,7 @@ Route::middleware('auth')->group(function () {
 
     Route::prefix('knowledgecafe')->namespace('KnowledgeCafe')->group(function () {
         Route::get('/', 'KnowledgeCafeController@index');
-        
+
         Route::prefix('library')->namespace('Library')->group(function () {
             Route::resource('books', 'BookController')
                 ->names([ 'index' => 'books.index', 'create' => 'books.create', 'show' => 'books.show', 'store' => 'books.store']);
@@ -73,7 +73,8 @@ Route::middleware('auth')->group(function () {
             ->only(['index', 'store', 'update', 'destroy'])
             ->names(['index' => 'books.category.index']);
         });
-        
+
+       Route::resource('weeklydoses', 'WeeklyDoseController')->only(['index']);        
     });
 
 });
